1. update kafka version 2. update jackson version 3. replace creating topics with more recent AdminClient

+++ kafka-injector-core/src/main/java/com/mikerusoft/kafka/injector/core/properties/Topic.java
@@ -16,5 +16,5 @@ public class Topic {
     private Generator[] valueGenerators;
     private Generator keyGenerator;
     private Integer partitions=1;
-    private Integer replicas=1;
+    private Short replicas=1;
 }
